// src/http-divida-ativa/models/parcelaOpcaoResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * api-imoveis
 * Cadastro local de inscricoes imobiliarias e servicos de IPTU / Divida Ativa.
 * OpenAPI spec version: 1.0.0
 */

export interface ParcelaOpcaoResponse {
  qtdeParcelas?: number
  valor1aParcela?: string
  valorJuros?: string
  valorTotalDescontoPrinc?: string
  valorTotalDescontoHonor?: string
  valorDescontos?: string
}